Use definition from mergedTypes
The value of the types in each schemas definition.PossibleTypes field
wouldn't have the merged version of the type.

So if one of the concrete types for an interface was a boundary type,
the possibleTypes map may not contain the merged version of the type
definition.

If the schema tree is processed later it could be missing the complete
type in the possible types field.

Use the mergedTypes value instead of the value from one of the source
trees.